pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atol-canopy"
version = "0.1.0"
description = "FastAPI metadata tracking system for the Australian Tree of Life (AToL) project."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "GPL-3.0-or-later" }
authors = [{ name = "AToL team" }]
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.23.2",
  "sqlalchemy>=2.0.22",
  "pydantic>=2.4.2",
  "pydantic-settings>=2.0.3",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg2-binary>=2.9.9",
  "alembic>=1.12.0",
  "python-multipart>=0.0.6",
  "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=7.4.0",
  "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
  "httpx>=0.27.0",
  "pytest>=7.4.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=7.0.0",
]
